[RW-373][374] Fix saving disasters as archived, fix redirection after creating terms and reorganize code for job/training status update based on user posting rights
