+++ html/modules/custom/reliefweb_entities/src/Entity/Training.php
@@ -2,11 +2,14 @@
 
 namespace Drupal\reliefweb_entities\Entity;
 
+use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\node\Entity\Node;
 use Drupal\reliefweb_entities\BundleEntityInterface;
 use Drupal\reliefweb_entities\DocumentInterface;
 use Drupal\reliefweb_entities\DocumentTrait;
+use Drupal\reliefweb_entities\OpportunityDocumentInterface;
+use Drupal\reliefweb_entities\OpportunityDocumentTrait;
 use Drupal\reliefweb_moderation\EntityModeratedInterface;
 use Drupal\reliefweb_moderation\EntityModeratedTrait;
 use Drupal\reliefweb_revisions\EntityRevisionedInterface;
@@ -17,11 +20,12 @@ use Drupal\reliefweb_utility\Helpers\UrlHelper;
 /**
  * Bundle class for training nodes.
  */
-class Training extends Node implements BundleEntityInterface, EntityModeratedInterface, EntityRevisionedInterface, DocumentInterface {
+class Training extends Node implements BundleEntityInterface, EntityModeratedInterface, EntityRevisionedInterface, DocumentInterface, OpportunityDocumentInterface {
 
   use DocumentTrait;
   use EntityModeratedTrait;
   use EntityRevisionedTrait;
+  use OpportunityDocumentTrait;
   use StringTranslationTrait;
 
   /**
@@ -91,6 +95,16 @@ class Training extends Node implements BundleEntityInterface, EntityModeratedInt
       'training_language' => $this->getEntityMetaFromField('training_language', 'TL'),
       'cost' => $cost,
     ];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function preSave(EntityStorageInterface $storage) {
+    // Update the entity status based on the user posting rights.
+    $this->updateModerationStatusFromPostingRights();
+
+    parent::preSave($storage);
   }
 
 }
