+++ html/modules/custom/reliefweb_entities/src/OpportunityDocumentInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\reliefweb_entities;
+
+/**
+ * Interface for the opportunity document entities like jobs and training.
+ */
+interface OpportunityDocumentInterface {
+
+}
